+++ tickets/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "ts-node-dev src/server.ts",
     "test": "jest --watchAll --no-cache",
-    "test:ci":"jest"
+    "test:ci": "jest"
   },
   "jest": {
     "preset": "ts-jest",
@@ -30,6 +30,7 @@
     "node-nats-streaming": "^0.3.2",
     "nodemailer": "^6.9.16",
     "pg": "^8.13.1",
+    "pg-mem": "^3.0.5",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.7.3"
   },
